extracted context menu to reusable component
